chore(os): update oss to use nodejs construct
