--- src/cdk-lib/opensearch-vectorindex/vector-index/index.js
@@ -1,1 +1,0 @@
-"use strict";
